Adapt PriceCardBody to work as promo
Adapated the PriceCardBody to conditionally render as a promo card or regular card. This functionality is triggered through the isPromo prop, which is passed down through the PriceCardFeatures, that renders different fontSizes and spacing for a promo specific card. Promo card will always be position 2 or data[1] in the grid. However, should rewrite this because the code isn't that DRY and there is some prop-drilling going on.